improved logging in Driver. changed all socket connections to 32 bit addresses - ***** note this will break blender connections until they are also changed.
git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@542 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/caviar/eventio/AEMulticastInput.java
@@ -127,7 +127,7 @@ public class AEMulticastInput extends Thread {
 //            short[] addr=packet.getAddresses();
 //            int[] ts=packet.getTimestamps();
             for(int i=0;i<nEventsInPacket;i++){
-                eventRaw.address=dis.readShort();
+                eventRaw.address=dis.readInt();
                 eventRaw.timestamp=dis.readInt();
                 packet.addEvent(eventRaw);
 //                addr[i]=dis.readShort();
